LayoutTests/inspector-obsolete/timeline/timeline-network-received-data-expected.txt

CONSOLE MESSAGE: line 7: Script evaluated
Tests the Timeline API instrumentation of a network resource received data

Script resource loaded
Resource received data has length, test passed.
